+++ etherparse/tests/internet/mod.rs
@@ -1,6 +1,4 @@
 pub mod ip;
 pub mod ipv4_extensions;
-pub mod ipv6;
 pub mod ipv6_extensions;
 pub mod ipv6_fragment;
-pub mod ipv6_raw_extension;